Refactor layout to get ready for activity desktop view
